mirofish_poc/requirements.txt
langgraph>=0.0.35
langchain>=0.1.13
langchain-google-genai>=1.0.1
pydantic>=2.6.4
python-dotenv>=1.0.1
